+++ src/io/trr_io.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 
 use crate::errors::{ReadXdrError, WriteXdrError, XdrError};
 use crate::io::xdrfile::{
-    self, OpenMode, XdrFile, XdrGroupWriter, XdrRangeReader, XdrReader, XdrWriter,
+    self, OpenMode, XdrFile, XdrFrameData, XdrGroupWriter, XdrRangeReader, XdrReader, XdrWriter,
 };
 use crate::iterators::AtomIterator;
 use crate::structures::{group::Group, vector3d::Vector3D};
@@ -55,6 +55,63 @@ impl<'a> XdrReader<'a> for TrrReader<'a> {
             phantom: PhantomData,
         })
     }
+
+    /// Transform `TrrReader` iterator into `XdrRangeReader<TrrReader>` iterator with the specified time range.
+    /// Time (`start_time` and `end_time`) must be provided in picoseconds.
+    ///
+    /// ## Details
+    /// The frames with time lower than `start_time` are skipped over, i.e., the properties of the atoms
+    /// are not read at all, making the `XdrRangeReader` very efficient. Note however that calling this function nonetheless
+    /// involves some initial overhead, as it needs to locate the starting point of the iteration.
+    ///
+    /// Iteration is ended at the frame with time corresponding to `end_time` or once the end of the trr file is reached.
+    /// The range is inclusive on both ends, i.e., frames with `time = start_time` and `time = end_time` will be included in the iteration.
+    ///
+    /// ## Returns
+    /// `XdrRangeReader<TrrReader>` if the the specified time range is valid. Else returns `ReadXdrError`.
+    ///
+    /// ## Example
+    /// Creating an `XdrRangeReader` iterator over a trr file.
+    /// ```no_run
+    /// use groan_rs::prelude::*;
+    ///
+    /// // load system from file
+    /// let mut system = System::from_file("system.gro").unwrap();
+    ///
+    /// // create the iterator for partial reading of a trr file
+    /// let range_iterator = system
+    ///     .trr_iter("trajectory.trr")
+    ///     .unwrap()
+    ///     .with_range(10_000.0, 100_000.0)
+    ///     .unwrap();
+    /// ```
+    ///
+    /// ## Notes
+    /// - If the frame corresponding to the `start_time` doesn't exist in the trr file,
+    /// the iterator starts at the frame closest in time to but greater than the `start_time`.
+    /// - If either the `start_time` or the `end_time` is negative, it results in a `ReadXdrError::TimeRangeNegative` error.
+    /// - If the `start_time` is greater than the `end_time`, it results in a `ReadXdrError::InvalidTimeRange` error.
+    /// - If the `start_time` exceeds the time of any frame in the xtc file, it results in a `ReadXdrError::StartNotFound` error.
+    fn with_range(
+        self,
+        start_time: f32,
+        end_time: f32,
+    ) -> Result<XdrRangeReader<'a, TrrReader<'a>>, ReadXdrError> {
+        xdrfile::sanity_check_timerange(start_time, end_time)?;
+
+        let reader = XdrRangeReader::new(self, start_time, end_time);
+
+        // jump to the start of iteration
+        unsafe {
+            if xdrfile::trr_jump_to_start(reader.xdrreader.trr.handle, reader.start_time as c_float)
+                != 0 as c_int
+            {
+                return Err(ReadXdrError::StartNotFound(start_time.to_string()));
+            }
+        }
+
+        Ok(reader)
+    }
 }
 
 impl<'a> Iterator for TrrReader<'a> {
@@ -80,83 +137,23 @@ impl<'a> Iterator for TrrReader<'a> {
     }
 }
 
-impl<'a> TrrReader<'a> {
-    /// Transform `TrrReader` into `TrrRangeReader` with the specified time range.
-    /// ## Warning
-    /// This does not create a valid `TrrRangeReader`! This is a helper function.
-    /// Use `TrrRangeReader::new()` to create a valid `TrrRangeReader`.
-    fn with_range(self, start_time: f32, end_time: f32) -> TrrRangeReader<'a> {
-        TrrRangeReader {
-            system: self.system,
-            trr: self.trr,
-            phantom: self.phantom,
-            start_time,
-            end_time,
-        }
-    }
-}
-
-/// Iterator over a time range of a trr file.
-pub struct TrrRangeReader<'a> {
-    system: *mut System,
-    trr: XdrFile,
-    phantom: PhantomData<&'a mut System>,
-    /// starting time of the iteration (in ps)
-    start_time: f32,
-    /// ending time of the iteration (in ps)
-    end_time: f32,
-}
-
-impl<'a> XdrRangeReader<'a> for TrrRangeReader<'a> {
-    /// Open a trr file and create an iterator for a specified range of frames.
-    /// The `start_time` and `end_time` parameters should be provided in picoseconds.
-    ///
-    /// ## Returns
-    /// - `TrrRangeReader` iterator for reading the trr file.
-    /// - `ReadXdrError` if the iterator couldn't be constructed.)
-    fn new(
-        system: &'a mut System,
-        filename: impl AsRef<Path>,
-        start_time: f32,
-        end_time: f32,
-    ) -> Result<Self, ReadXdrError> {
-        xdrfile::sanity_check_timerange(start_time, end_time)?;
-
-        let reader = TrrReader::new(system, filename)?.with_range(start_time, end_time);
-
-        // jump to the start of iteration
-        unsafe {
-            if xdrfile::trr_jump_to_start(reader.trr.handle, reader.start_time as c_float)
-                != 0 as c_int
-            {
-                return Err(ReadXdrError::StartNotFound(start_time.to_string()));
-            }
-        }
-
-        Ok(reader)
-    }
-}
-
-impl<'a> Iterator for TrrRangeReader<'a> {
+impl<'a> Iterator for XdrRangeReader<'a, TrrReader<'a>> {
     type Item = Result<&'a mut System, ReadXdrError>;
 
-    /// Read next frame in a trr file.
-    ///
-    /// ## Returns
-    /// `None` in case the specified time range of the frames has been fully read.
-    /// `Some(&mut System)` in case the frame has been read successfully.
-    /// `Some(ReadXdrError)` in case an error occured while reading.
+    /// Read the next frame in the specified time range of the trr file.
     fn next(&mut self) -> Option<Self::Item> {
         unsafe {
-            match TrrFrameData::from_frame(&mut self.trr, (*self.system).get_n_atoms()) {
+            let system = self.xdrreader.system;
+
+            match TrrFrameData::from_frame(&mut self.xdrreader.trr, (*system).get_n_atoms()) {
                 None => None,
                 Some(Err(e)) => Some(Err(e)),
                 Some(Ok(data)) => {
-                    if (*self.system).get_simulation_time() >= self.end_time {
+                    if (*system).get_simulation_time() >= self.end_time {
                         None
                     } else {
-                        data.update_system(&mut *self.system);
-                        Some(Ok(&mut *self.system))
+                        data.update_system(&mut *system);
+                        Some(Ok(&mut *system))
                     }
                 }
             }
@@ -172,7 +169,7 @@ impl System {
     /// `TrrReader` if the trr file exists and matches the structure file.
     /// Else returns `ReadXdrError`.
     ///
-    /// ## Example
+    /// ## Examples
     /// Iterating through a trr trajectory and calculating
     /// and printing the current center of geometry of the system.
     ///
@@ -186,6 +183,29 @@ impl System {
     ///
     ///     // loop through the trajectory
     ///     for raw_frame in system.trr_iter("trajectory.trr")? {
+    ///         let frame = raw_frame?;
+    ///         println!("{:?}", frame.group_get_center("all"));
+    ///     }
+    ///
+    ///     Ok(())
+    /// }
+    /// ```
+    ///
+    /// You can also iterate over just a part of the trajectory.
+    /// Here, only frames in the time range 10-100 ns will be read.
+    /// **Do not use the `skip` method for skipping over the initial frames
+    /// of the trajectory as that is very inefficient.**
+    /// ```no_run
+    /// use groan_rs::prelude::*;
+    /// use groan_rs::errors::ReadXdrError;
+    ///
+    /// fn example_fn() -> Result<(), ReadXdrError> {
+    ///     let mut system = System::from_file("system.gro").unwrap();
+    ///
+    ///     for raw_frame in system
+    ///         .trr_iter("trajectory.trr")?
+    ///         .with_range(10_000.0, 100_000.0)?
+    ///     {
     ///         let frame = raw_frame?;
     ///         println!("{:?}", frame.group_get_center("all"));
     ///     }
@@ -261,8 +281,8 @@ impl System {
         filename: impl AsRef<Path>,
         start_time: f32,
         end_time: f32,
-    ) -> Result<TrrRangeReader, ReadXdrError> {
-        TrrRangeReader::new(self, filename, start_time, end_time)
+    ) -> Result<XdrRangeReader<TrrReader>, ReadXdrError> {
+        TrrReader::new(self, filename)?.with_range(start_time, end_time)
     }
 }
 
@@ -567,7 +587,7 @@ struct TrrFrameData {
     forces: Vec<[f32; 3]>,
 }
 
-impl TrrFrameData {
+impl XdrFrameData for TrrFrameData {
     /// Read `TrrFrameData` from a trr frame.
     ///
     /// ## Returns